Update GitHub Actions workflow for better artifact handling
The workflow was updated to use a different version of the 'actions/checkout' step. Furthermore, the process of renaming and uploading executables for Windows, MacOS and Linux was split into separate steps, providing a cleaner approach in our CI/CD pipeline. This should improve the upload and artifact handling for our builds on different OS.