Update GitHub workflow and version.json
Modified the package.yml file in GitHub workflow to account for specific actions when operating system is Windows. Also, updated the versions.json file to include details for version 3.1.0. The changes in package.yml specifically enable setting the version number and the upload URL differently for Windows and non-Windows systems.